HOMME, 3D SL: New 3D QLT approach.
Previously, constructed a tree over the 3D volume. I discovered that
is not a good approach. The key is that each level's global mass is
only slightly different than at the start of the time step (due to
error in the reconstruction of levels) and should be perturbed only
enough to fix it. A 3D tree with ISL perturbs it more b/c of the
overall mass fix that is needed.

In the new approach, insert a function at the root of the QLT tree
that assigns each level its target mass based on the above
observation. Later, I will need to do some s/w redesign to permit
clean injection of such a function.