Homme(xx)/SL: Modify vertical treatment in enhanced trajectory method.
The atmosphere semi-Lagrangian tracer transport method's enhanced trajectory
method (ETM) is a stealth feature intended to make tracer transport time steps
even longer and, in any case, more flexible.

This PR modifies the vertical discretization relative to the original ETM
implementation, in light of issues at and around the tropopause when using the
ETM in test v3 simulations. This new approach resolves the tropopause issue.

In addition to revising the ETM implementation, this PR strengthens the
homme_integration test suite as follows:
* add a new planar tracer transport test with analytical solution at the end
  time, with space- and time-varying surface pressure;
* add instances of this test to the homme_integration test suite;
* add a cmake tool to check error output for the recently added sphere and new
  plane tracer transport tests when running this suite;
* activate the Kokkos builds in the non-BFB part of homme_integration, turning
  on the parts of the unit tests that are meaningful in the non-BFB build and
  the recently added and new tracer transport tests.